+++ packages/categories/collections/cats.js
@@ -26,6 +26,19 @@ K.extend({
 			sort: { name:1 }
 		});
 	},
+	findCatsByType: function(type) {
+		var w = {
+				active: true
+			};
+
+		if(_.isString(type))
+			w.type = type;
+
+		return Categories.find(w, {
+				sort: { rank:-1, name:1 },
+				limit: 50
+			});
+	},	
 	findCatsByName: function(initial, type) {
 		
 		if(!initial) {
@@ -38,9 +51,9 @@ K.extend({
 				w.type = type;
 			
 			return Categories.find(w, {
-				sort: { name:1 },
-				limit: 50
-			});
+					sort: { name:1 },
+					limit: 50
+				});
 		}
 
 		initial = K.Util.sanitize.regExp(initial);
@@ -50,7 +63,7 @@ K.extend({
 
 		var ex = new RegExp('^'+ initial, 'i');
 			
-		var curCat = Categories.find({
+		return Categories.find({
 				active: true,
 				name: ex,
 				type: type
@@ -58,7 +71,5 @@ K.extend({
 				sort: { name:1 },
 				limit: 50
 			});
-
-		return curCat;
 	}
 });